Bugfix: _lastSyncRequestTickTime should be of type 'unsigned long' as that's the same type as 'millis()' returns, otherwise an overrun could occur earlier than necessary.

+++ src/ArduinoIoTCloudTCP.h
@@ -96,7 +96,7 @@ class ArduinoIoTCloudTCP: public ArduinoIoTCloudClass
 
     State _state;
 
-    int _lastSyncRequestTickTime;
+    unsigned long _lastSyncRequestTickTime;
     String _brokerAddress;
     uint16_t _brokerPort;
     uint8_t _mqtt_data_buf[MQTT_TRANSMIT_BUFFER_SIZE];
